Move to new subdirectory in Analysis called Presburger
