+++ mlir/include/mlir/Analysis/Presburger/Matrix.h
@@ -2,6 +2,7 @@
 #define MLIR_ANALYSIS_MATRIX_H
 
 #include <llvm/Support/raw_ostream.h>
+#include <llvm/ADT/SmallVector.h>
 
 #include <cassert>
 #include <iomanip>
