+++ mlir/lib/Analysis/Presburger/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_mlir_library(MLIRPresburger
+  Simplex.cpp
+  )
+
+target_link_libraries(MLIRPresburger
+  PUBLIC
+  MLIRAnalysis
+  )
+
