+++ mlir/unittests/Analysis/CMakeLists.txt
@@ -1,7 +1,1 @@
-add_mlir_unittest(MLIRAnalysisTests
-  MatrixTest.cpp
-  SimplexTest.cpp
-)
-
-target_link_libraries(MLIRAnalysisTests
-  PRIVATE MLIRAnalysis)
+add_subdirectory(Presburger)
